Move GraphQL `typeDefs` to its own file
The code was refactored to create a file for the GraphQL typeDefs and
TypeScript interfaces.
