+++ src/server.ts
@@ -5,49 +5,13 @@ import { PrismaClient } from '@prisma/client';
 import { startStandaloneServer } from '@apollo/server/standalone';
 import { PrismaClientKnownRequestError } from '@prisma/client/runtime/library';
 
+import { typeDefs, User, UserInput } from './typedefs.js';
+
 export let prisma: PrismaClient;
 
 export const initializeDatabaseInstance = (): void => {
   prisma = new PrismaClient();
 };
-
-const typeDefs = `#graphql
-  type Query {
-    hello: String!
-  }
-
-  input UserInput {
-    name: String!
-    email: String!
-    password: String!
-    birthDate: String!
-  }
-
-  type User {
-    id: ID!
-    name: String!
-    email: String!
-    birthDate: String!
-  }
-
-  type Mutation {
-    createUser(user: UserInput!): User
-  }
-`;
-
-export interface UserInput {
-  name: string;
-  email: string;
-  password: string;
-  birthDate: string;
-}
-
-export interface User {
-  id: number;
-  name: string;
-  email: string;
-  birthDate: Date;
-}
 
 class ServerErrorGQL extends GraphQLError {
   public code: number;
