+++ src/typedefs.ts
@@ -1,0 +1,37 @@
+export const typeDefs = `#graphql
+  type Query {
+    hello: String!
+  }
+
+  input UserInput {
+    name: String!
+    email: String!
+    password: String!
+    birthDate: String!
+  }
+
+  type User {
+    id: ID!
+    name: String!
+    email: String!
+    birthDate: String!
+  }
+
+  type Mutation {
+    createUser(user: UserInput!): User
+  }
+`;
+
+export interface UserInput {
+  name: string;
+  email: string;
+  password: string;
+  birthDate: string;
+}
+
+export interface User {
+  id: number;
+  name: string;
+  email: string;
+  birthDate: Date;
+}
